refactor: reorganize module structure
