+++ testcontainers/Cargo.toml
@@ -18,7 +18,6 @@ conquer-once = { version = "0.4", optional = true }
 dirs = "5.0.1"
 futures = "0.3"
 log = "0.4"
-rand = "0.8"
 serde = { version = "1", features = ["derive"] }
 serde-java-properties = "0.1.1"
 serde_json = "1"
