+++ testcontainers/src/core.rs
@@ -1,20 +1,13 @@
-pub use self::{
-    container_async::ContainerAsync,
-    image::{ContainerState, ExecCommand, Host, Image, ImageArgs, Port, RunnableImage, WaitFor},
+pub use self::image::{
+    ContainerState, ExecCommand, Host, Image, ImageArgs, Port, RunnableImage, WaitFor,
 };
 
-#[cfg(feature = "blocking")]
-pub use self::container::Container;
+pub use self::containers::*;
 
-#[cfg(feature = "blocking")]
-mod container;
-mod container_async;
 mod image;
 
-/// Helper traits to start containers.
-pub mod runners;
-
 pub(crate) mod client;
+pub(crate) mod containers;
 pub(crate) mod env;
 pub(crate) mod logs;
 pub(crate) mod network;
