+++ testcontainers/src/core/containers/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod async_container;
+#[cfg(feature = "blocking")]
+pub(crate) mod sync_container;
+
+pub use async_container::ContainerAsync;
+
+#[cfg(feature = "blocking")]
+#[cfg_attr(docsrs, doc(cfg(feature = "blocking")))]
+pub use sync_container::Container;
