+++ testcontainers/src/images/mod.rs
