+++ testcontainers/tests/dual_stack_host_ports.rs
@@ -2,10 +2,7 @@
 
 use std::net::{Ipv6Addr, TcpListener};
 
-use testcontainers::{
-    core::{runners::SyncRunner, WaitFor},
-    GenericImage,
-};
+use testcontainers::{core::WaitFor, runners::SyncRunner, GenericImage};
 
 /// Test the functionality of exposing container ports over both IPv4 and IPv6.
 #[test]
